Recalc.Engine

Description

The evaluation strategy is inspired by the paper "Build systems à la carte", it maintains a DocumentStore that tracks types, values and meta data (dummied out for now) for all cells, and a dependency graph of the cells for checking which cells are dirty (need to be recomputed).

References: Andrey Mokhov, Neil Mitchell and Simon Peyton Jones. Build Systems à la Carte. Proceedings of the ACM on Programming Languages, Volume 2, Issue ICFP.

Engine Operations

type Inputs = [(CellRef, (Maybe (String, CellType), Meta))]

inputs are given by the location, maybe an input (and its type), and meta data

type CycleOf t = [(CellRef, Cell (ErrorOf t) t (ValueOf t))]

when there is a dependency cycle, the cycle's locations are returned together with the data stored there

type ResultsOf t = [(CellRef, Cell (ErrorOf t) t (ValueOf t))]

the results are locations and the data stored there

recalc :: forall t. Recalc t => Inputs -> EngineStateOf t -> (Either (CycleOf t) (ResultsOf t), EngineStateOf t)

recalculate the inputs and everything that depends on those (transitively), or return the cycle if there is a dependency cycle.

recalcAll :: forall t. Recalc t => EngineStateOf t -> (Either (CycleOf t) (ResultsOf t), EngineStateOf t)

recalculate the whole sheet (marking everything as dirty), or return the cycle if there is a dependency cycle.

parseTerm :: Recalc t => CellRef -> (String, CellType) -> Parsed t

parse a cell (depends on the cell's location) returns formula or value (both as term)

Spreadsheet Language Interface

class (Pretty t, Pretty (ValueOf t)) => Recalc t where

The spreadsheet language interface. An instantiation of it can be used to declare the semantics of the provided language in a spreadsheet environment.

Associated Types

type EnvOf t

a custom environment (e.g. to bind global variables)

type ErrorOf t

a custom error type (errors during type inference, or evaluation)

type ValueOf t

the value t evaluates to (values are types!)

Methods

parseCell :: CellType -> ReaderT SheetId (Parsec Void String) t

the language may differentiate between cells parsed as formula and value

depsOf :: t -> Set CellRangeRef

specify how to compute the cell references a term depends on

infer :: t -> FetchOf t (ValueOf t)

specify how a term's type is inferred

eval :: t -> FetchOf t (ValueOf t)

specify how a term is evaluated

Fetch Monad

data Fetch env err v a

Fetch callbacks can fail (using MonadError) and have access to other cells (see fetchType, fetchValue), and the custom context (using MonadReader)

type FetchOf t a = Fetch (EnvOf t) (ErrorOf t) (ValueOf t) a

data FetchError err

Evaluation of cells can always fail due to invalid formulas or refs

Constructors

InvalidFormula ParseError

RefError

SemanticError err

Actions

fetchType :: CellRef -> Fetch env err v v

fetch the type of the value at a cell reference

fetchValue :: CellRef -> Fetch env err v v

fetch the value stored at a cell reference

throwSemanticError :: err -> Fetch env err v a

fail with a user-defined error

Run Action (for debugging purposes)

runFetchWith

Arguments

:: env

-> (Kind -> CellRef -> Except (FetchError err) v)
callback that returns a value/type (depending on its kind)

-> Fetch env err v a
a fetch task to run

-> Either (FetchError err) a

Engine State

type EngineStateOf t = EngineState (EnvOf t) (ErrorOf t) t (ValueOf t)

newEngineState :: env -> EngineState env err t v

create a new engine state. the engine state is completely empty, no handles on any documents.

engineEnv :: EngineState env err t v -> env

engineDocs :: EngineState env err t v -> DocumentStore err t v

State Updates

mapEnv :: (env -> env) -> EngineState env err t v -> EngineState env err t v

modify the custom environment

mapDocs :: (DocumentStore err t v -> DocumentStore err t v) -> EngineState env err t v -> EngineState env err t v

modify the document store

deleteSheetId :: SheetId -> EngineState env err t v -> EngineState env err t v

delete a whole sheet by its id from the dependency graph (useful when renaming a sheet)

Types

type DocumentStoreOf t = DocumentStore (ErrorOf t) t (ValueOf t)

data Document err t v

Constructors

Document

Fields

sheetOrder :: ![Text]

sheets may be re-ordered (visually)

sheets :: !(Map SheetName (Sheet err t v))

store of sheets by their name

type SheetOf t = Sheet (ErrorOf t) t (ValueOf t)

data Cell err t v

Constructors

Cell

Fields

cell :: !(Maybe ((String, CellType), Maybe (t, Maybe (v, Maybe v))))

a cell maybe only meta data. when it has a term (and therefore the input and celltype are available), it can have a type. if it has a type, it can have a value.

cellDeps :: !(Set CellRangeRef)

dependencies of the cell

cellError :: !(Maybe (FetchError err))

when the cell encounters an error (during inference or evaluation)

cellMeta :: !Meta

meta data (currently useless, should include style info like it once did)

data CellType

Constructors

CellFormula
cell formulas (starting with =)

CellValue
regular values

data Meta

Constructors

Meta

re-export core definitions

type SheetName = Text

type SheetId = (URI, SheetName)

type CellAddr = (Int, Int)

row and column

row :: CellAddr -> Int

column :: CellAddr -> Int

type CellRange = (CellAddr, CellAddr)

type CellRangeRef = (SheetId, CellRange)

type CellRef = (SheetId, CellAddr)

readExcel :: Text -> Maybe CellAddr

read an spreadsheet address of the form column-row (columns are labelled "A..Z, AA..", and rows enumerated)

>>> readExcel "A2"
Just (1,0)

showExcel26 :: CellAddr -> String

show a zero-indexed cell address in Excel-style

>>> showExcel (0,4)
"E1"