Recalc.EngineSpec

Description

Implements a simple calculator language and uses it instantiate the Recalc interface. Use this language to validate the spreadsheet engine.

Documentation

type Result = Either (FetchError ())

run :: [[(CellAddr, Maybe String)]] -> (ResultsOf Term, EngineStateOf Term)

evalSheet all inputs in sequence

evalSheet :: [[(CellAddr, Maybe String)]] -> [(CellAddr, Int)]

evalSheet inputs in sequence and prepare Results for hspec

evalSheetKeepErrors :: [[(CellAddr, Maybe String)]] -> [(CellAddr, Result Int)]

spec :: Spec

the specification checks a few (single-document, single-sheet) examples to make sure results are expected (incl. making sure dependencies are resolved)

data SheetRef

when referencing a cell, the uri and sheetName can be omited (defaults to "current" sheet)

Constructors

Explicit

Fields

sheetId :: SheetId

Implicit

Fields

sheetId :: SheetId

data Term

a simple term language

Constructors

Num Int

Add Term Term

Ref SheetRef CellAddr

Sum SheetRef CellRange

Funky

type Parser = ReaderT SheetId (Parsec Void String)

formulaP :: Parser Term

simple parser, does not deal with lexing (whitespaces might break things)

valueP :: Parser Term

Orphan instances

IsString (Maybe String)
use "" for cell deletion and others for new input

Instance details

Methods

fromString :: String -> Maybe String