Recalc.Language

Description

Defines the values for the lambda calculus, evaluation and bi-directional type checking. It instantiates the classes Recalc and UniverRecalc such that the language can be used by the Univer frontend.

The original implementation of the core calculus is based on "A Tutorial Implementation of a Dependently Typed Lambda Calculus".

This extension lifts the calculus to spreadsheets and adds homogeneous tensor types which can be constructed using cell-references.

References: Andres Löh, Conor McBride, Wouter Swierstra. A Tutorial Implementation of a Dependently Typed Lambda Calculus. Fundamenta Informaticae, Volume 102, Issue 2.

Types

newtype Env

the environment contains just the globally bound variables

Constructors

Env

Fields

globals :: Globals

data Term (m :: Mode)

data Mode

modes for typing: bi-directional type checking

Constructors

Infer
inferrable terms

type Type = Value

data SemanticError

all semantic errors that can occur during type checking or evaluation

Constructors

TypeMismatch Type Type

TypeMismatchPi (Term Check) Type

TypeOfBound

ExpectedInt Value

ExpectedSet (Term Check)

IllegalApplication Type

IllegalImplicitApplication Type

IllegalArrayType [(CellAddr, Type)]

UnknownIdentifier Name

UnificationError UnificationError

Initial Global Variables

prelude :: Globals

for test only

check' :: [Type] -> Type -> Term Check -> Spreadsheet (Term Check)

fails if the type does not check, returns an elaborated term otherwise

eval' :: [Value] -> Term m -> Spreadsheet Value

infer' :: [Type] -> Term Infer -> Spreadsheet (Type, Term Infer)

data Decl

a global variable must declare its Type but not necessarily its Value

Constructors

Decl

Fields

declType :: Type

declValue :: Maybe Value

data Value

terms evaluate to values. values with bound (such as VPi) variables in them are represented using a HOAS.

note that the return type is a monadic value such that those values can easily resolve types/values from other cells.

Constructors

VLam !Arg !(Maybe CaseInsensitive) !(Value -> Spreadsheet Value)
lambda abstractions

VSet !Int
set of k-th order (hierarchy of types)

VPi !Arg !(Maybe CaseInsensitive) !Value !(Value -> Spreadsheet Value)
dependent function

VLit Lit
type literal

VLitOf LitOf
value literal

VTensor VTensorDescriptor
tensor type

VTensorOf VTensorDescriptor (Array Value)
tensor value

VNeutral !Neutral
stuck terms

data VTensorDescriptor

Constructors

VTensorDescriptor Value [Value]

data Neutral

Constructors

NFree !Name

NRef !SheetId !CellRange

NApp !Arg !Neutral !Value

type Globals = Map Name Decl

vapp :: Value -> Value -> Spreadsheet Value

application (must be able to fail)

vbool :: Type

type literals

vboolOf :: Bool -> Value

boolean values

vfree :: Name -> Value

free variable

vfun :: Value -> Value -> Value infixr 9

function type constructor (no bound type)

vlam :: Arg -> Maybe CaseInsensitive -> (Value -> Value) -> Value

vpi :: Arg -> Maybe CaseInsensitive -> Value -> (Value -> Value) -> Value

dependent function

Orphan instances

Recalc (Term 'Infer)
instantiate the interface with the language defined above

Instance details

Associated Types

type EnvOf (Term 'Infer)

type ErrorOf (Term 'Infer)

type ElaborationOf (Term 'Infer)

type TypeOf (Term 'Infer)

type ValueOf (Term 'Infer)

Methods

parseCell :: CellType -> ReaderT SheetId (Parsec Void String) (Term 'Infer)

depsOf :: Term 'Infer -> Set CellRangeRef

infer :: Term 'Infer -> FetchOf (Term 'Infer) (TypeOf (Term 'Infer))

inferElaborate :: Term 'Infer -> FetchOf (Term 'Infer) (TypeOf (Term 'Infer), ElaborationOf (Term 'Infer))

eval :: ElaborationOf (Term 'Infer) -> FetchOf (Term 'Infer) (ValueOf (Term 'Infer))

UniverRecalc (Term 'Infer)
for error annotations and sheet-defined functions (not implemented) we need to instantiate the Univer-specific interface too:

Instance details

Methods

errorAnnotation :: ErrorOf (Term 'Infer) -> Annotation

define :: Text -> Map CellAddr (Maybe ((String, CellType), Maybe (Term 'Infer, Maybe ((TypeOf (Term 'Infer), ElaborationOf (Term 'Infer)), Maybe (ValueOf (Term 'Infer)))))) -> [(Text, CellRange)] -> CellRange -> EnvOf (Term 'Infer) -> Either (ErrorOf (Term 'Infer)) (EnvOf (Term 'Infer))