Recalc.Repl

Description

Provides a few utilities to experiment with the core language to examine outcomes of parsing, type checking and inference, evaluation and sheet recomputation.

Documentation

data ExampleData

Constructors

ExampleData

Fields

exampleData'f, exampleData'v :: Nullable Text

type Result = Either (FetchError SemanticError)

uri :: URI

parseFormula :: String -> Result (Term Infer)

parseValue :: String -> Result Value

check :: [(Name, Decl)] -> Type -> Term Check -> Result ()

infer :: [(Name, Decl)] -> Term Infer -> Result Type

eval :: [(Name, Decl)] -> Term m -> Result Value

Engine

type EngineState = EngineState () () ExampleData (Term Infer)

newEngineState :: EngineState

runRecompute :: [(CellAddr, String)] -> EngineState -> (Either [CellAddr] NewValues, EngineState)

run a recomputation of a single spreadsheet

evalRecompute :: [(CellAddr, String)] -> EngineState -> Either [CellAddr] NewValues

run recomputation but only return result

re-export

pretty :: Pretty a => a -> Doc ann

>>> pretty 1 <+> pretty "hello" <+> pretty 1.234
1 hello 1.234

Orphan instances

(Pretty a, Pretty b) => Pretty (Either a b)

Instance details

Methods

pretty :: Either a b -> Doc ann

prettyList :: [Either a b] -> Doc ann