Recalc.Semantics

Description

Implementation of a very simple calculus based on "A tutorial implementation of a dependently typed lambda calculus".

This implementation implements the type level hierarchy soundly. It adds type assumptions and type/value assumptions for Boolean values and a few declarations to interact with them.

Documentation

data Decl

Constructors

Decl

Fields

declType :: Value

declValue :: Maybe Value

type Globals = Map Name Decl

data Env

Constructors

Env

Fields

sheetId :: SheetId

globals :: Globals

type Type = Value

values are types

data SemanticError

Constructors

TypeMismatch Type Type

TypeMismatchPi (Term Check) Type

TypeOfBound

ExpectedSet (Term Check)

IllegalApplication Type

IllegalArrayType [(CellAddr, Type)]

UnknownIdentifier Name

UnknownError Text

throwSemanticError :: MonadError (FetchError SemanticError) m => SemanticError -> m a

bindType :: Name -> Value -> Fetch (Term Infer) a -> Fetch (Term Infer) a

deps' :: Term m -> Set CellRange

assertInt :: Value -> Fetch (Term Infer) Int

eval' :: Term m -> Fetch (Term Infer) Value

concatMapM :: Monad m => (a -> m [b]) -> [a] -> m [b]

flattenTensor :: Value -> [Value]

check' :: Int -> Type -> Term Check -> Fetch (Term Infer) ()

infer' :: Int -> Term Infer -> Fetch (Term Infer) Type

inferLiteral :: LitOf -> Type

inferUniverse' :: Int -> Term Check -> Fetch (Term Infer) Int

Normal Forms

data Neutral

Constructors

NFree !Name

NRef !SheetId !CellRange

NApp !Neutral !Value

data VTensorDescriptor

Constructors

VTensorDescriptor Value [Int]

data Value

Constructors

VLam !(Maybe CaseInsensitive) !(Value -> Fetch (Term Infer) Value)

VSet !Int

VPi !(Maybe CaseInsensitive) !Value !(Value -> Fetch (Term Infer) Value)

VLit Lit

VLitOf LitOf

VTensor VTensorDescriptor

VTensorOf VTensorDescriptor (Array Value)

VNeutral !Neutral

vfree :: Name -> Value

vfun :: Value -> Value -> Value infixr 9

vapp :: Value -> Value -> Fetch (Term Infer) Value

vlam :: Maybe CaseInsensitive -> (Value -> Value) -> Value

vlams :: [Text] -> ([Value] -> Value) -> Value

vpi :: Maybe CaseInsensitive -> Value -> (Value -> Value) -> Value

vpis :: [(Text, Value)] -> ([Value] -> Value) -> Value

vbool :: Type

vint :: Type

vboolOf :: Bool -> Value

vintOf :: Int -> Value

vtensor :: Value -> [Int] -> Value

vtensorOf :: Value -> [Int] -> Array Value -> Value

cellRangeDescriptor :: Value -> CellRange -> VTensorDescriptor

cellRangeDescriptor' :: Value -> CellRange -> VTensorDescriptor

cellRangeAddrs :: CellRange -> [CellAddr]

Quoting

quote :: Value -> Term Check

valueP :: Parser Value

prelude :: Globals

treating Boolean values as globals

semanticErrorTitle :: SemanticError -> Text

Orphan instances

Language (Term 'Infer)
interface implementation

Instance details

Associated Types

type EnvOf (Term 'Infer)

type ErrorOf (Term 'Infer)

type ValueOf (Term 'Infer)

Methods

deps :: Term 'Infer -> Set CellRange

newEnv :: SheetId -> EnvOf (Term 'Infer)

infer :: Term 'Infer -> Fetch (EnvOf (Term 'Infer)) (ErrorOf (Term 'Infer)) (ValueOf (Term 'Infer)) (ValueOf (Term 'Infer))

inferValue :: EnvOf (Term 'Infer) -> ValueOf (Term 'Infer) -> Maybe (ValueOf (Term 'Infer))

eval :: Term 'Infer -> Fetch (EnvOf (Term 'Infer)) (ErrorOf (Term 'Infer)) (ValueOf (Term 'Infer)) (ValueOf (Term 'Infer))