Recalc.Server

Description

This module implements a reactor pattern for Recalc, ensuring asynchronous processing of spreadsheet updates and computations. A single threaded loop listens for JSON-RPC messages and maintains jobs.

Jobs are picked up by multiple "worker" threads such that long-running computations do not block messaging.

Example

type Protocol

type Api = ToApi Protocol

initialState = ...

main :: IO ()
main = runHandler Api initialState $ state ->
  hoist Api (runReaderT state) (namedHandlers server)
 where
  server :: Protocol (AsServerT (Handler EngineState))
  server = Protocol { ... }

Documentation

type Handler engine = ReaderT (State engine) IO

handler threads are in IO and have access to the state

liftEngine :: State engine a -> Handler engine a

debug :: Show a => String -> a -> Handler engine ()

dumpEngineState :: Show engine => Handler engine ()

scheduleJob :: ReaderT (State engine) IO () -> Handler engine ()

runHandler :: forall api engine. HasHandler api => engine -> (State engine -> HandlerT api IO) -> IO ()

reads lines from stdin and maintains a channel of reactor inputs with new requests, handler threads deal with requests by reading from the request channel.

type Id = Word64

data JsonRpcRequest params

Constructors

JsonRpcRequest

Fields

request'method :: String

request'id :: Maybe Id

request'params :: params

pattern JsonRpcNotification :: String -> params -> JsonRpcRequest params

data JsonRpc (sym :: Symbol) (params :: Type) (rsp :: Type)

class HasHandler api

Minimal complete definition

handle', hoist

type family HandlerT api (m :: Type -> Type)

handle :: forall api. HasHandler api => JsonRpcRequest Value -> Handler api -> Either String (IO ())

hoist :: HasHandler api => (forall x. m x -> n x) -> HandlerT api m -> HandlerT api n

class GenericMode mode

generic modes for protocol datatypes

Associated Types

type mode :- api :: Type infixl 0

data AsServerT (m :: Type -> Type)

type AsServer

Arguments

= AsServerT IO
the default server

data AsApi

type ToApi api' = GToSum (Rep (api' AsApi))

namedHandlers :: (GenericMode m, Generic (api' m), GSum (Rep (api' m))) => api' m -> GToSum (Rep (api' m))

generically converts an Api (AsServerT m) to its sum representation ep0 :| ep1 :| ...

data (path :: Symbol) :> (a :: Type) :: Type infixr 4

data a :<|> b infixr 3

Constructors

a :<|> b infixr 3

sendIO :: ToJSON a => a -> IO ()

send JSON-RPC message on stdout (locked to prevent interleaving)

data Nullable t

Constructors

Missing

Null

Is t

aesonOptions :: Options

JSON Instances