Recalc.Syntax.Arbitrary

Description

Implementation of the Arbitrary type class for terms, the arbitrary terms are fully random (they may not typecheck). A better approach in general would be to produce a restricted type and generate a term for it (à la Djinn).

Documentation

newtype SheetName

Constructors

SheetName Text

newtype Set0

Constructors

Set0 (Term Infer)

Orphan instances

Arbitrary URI

Instance details

Methods

arbitrary :: Gen URI

shrink :: URI -> [URI]

Arbitrary CaseInsensitive

Instance details

Methods

arbitrary :: Gen CaseInsensitive

shrink :: CaseInsensitive -> [CaseInsensitive]

CoArbitrary CaseInsensitive

Instance details

Methods

coarbitrary :: CaseInsensitive -> Gen b -> Gen b

Arbitrary (Term 'Check)

Instance details

Methods

arbitrary :: Gen (Term 'Check)

shrink :: Term 'Check -> [Term 'Check]

Arbitrary (Term 'Infer)

Instance details

Methods

arbitrary :: Gen (Term 'Infer)

shrink :: Term 'Infer -> [Term 'Infer]