Recalc.Syntax.Fixity

Description

This module provides builds pretty-printing utility functions and a parser for terms with operators.

Example Usage

import Recalc.Syntax.Fixity

data Op2 = Mul | Div | Add

data Term = Num Int | Op Op2 Term Term | ...

myOps :: [[Fixity Void Op2]]
myOps =
  [ [Infixl "*" Mul, Infixl "/" Div]
  , [Infixl "+" Add]
  ]

prettyTerm = pp 0
  where
    pp prec =
      let (left, right, ppOp1, ppRightOp2, ppLeftOp2, ppOp2) = ppHelpers myOps prec
       in case
        Num i -> pretty i
        Op op x y@(OP op' _ _)
          | op == op', right op -> ppRightOp2 pp op x y
        Op op x y
          | left op -> ppLeftOp2 pp op x y
          | otherwise -> ppOp2 pp op x y
        ...

termP :: Parser Term
termP = makeOpsParser
  (void . symbol)              -- parse operator from operator symbol
  (numP | parens termP)      -- parse atomic terms and parenthesized one recursively
  (error "no unary operators") -- skip (since there are no Prefix, or Postfix operators)
  Op                           -- build term from binary operator

Documentation

data Fixity op1 op2

Fixity declaration (unary prefix and postfix operators, as well as left-associative, right-associative and non-associative infix operators).

Constructors

Prefix String op1

Postfix String op1

Infixl String op2

Infixr String op2

Infix String op2

makeOpsParser

Arguments

:: MonadPlus f

=> (String -> f ())
operator symbol parser

-> f a
"atomic" term parser (must not accept ε)

-> (op1 -> a -> a)
constructor for unary operators

-> (op2 -> a -> a -> a)
constructor for binary operators

-> [[Fixity op1 op2]]
fixity declaration (precedence list)

-> f a

construct a parser from operator table

ppHelpers :: (Eq op1, Eq op2) => [[Fixity op1 op2]] -> Int -> (op2 -> Bool, op2 -> Bool, (Int -> t1 -> Doc ann) -> op1 -> t1 -> Doc ann, (Int -> t2 -> Doc ann) -> op2 -> t2 -> t2 -> Doc ann, (Int -> t3 -> Doc ann) -> op2 -> t3 -> t3 -> Doc ann, (Int -> t4 -> Doc ann) -> op2 -> t4 -> t4 -> Doc ann)

pretty-print terms minimally respecting specified operator precedences and associativities

sym :: Fixity op1 op2 -> String