Recalc.Syntax.Term

Description

Types definitions for a very basic dependently typed lambda calculus based on "A tutorial implementation of a dependently typed lambda calculus".

Terms

data Mode

modes for typing: bi-directional type checking

Constructors

Check
checkable terms (and elaborated terms)

Infer
inferrable terms

Names

newtype CaseInsensitive

Constructors

CaseInsensitive

Fields

originalText :: Text

caseInsensitive :: CaseInsensitive -> Text

data Name

Constructors

Global !CaseInsensitive

Quote !Int
bound variables during quoting

Implicit !Int
implicit variables inserted (unification variables)

data Arg

whether a binder is implicit or not

Constructors

EArg

IArg

pat :: Text -> Maybe CaseInsensitive

data RefInfo

cell references may be given [file_name.rc]sheet!C3, sheet!C3, or C3; corresponding to FullySpecified, SheetOnly, and Unspecified respectively

Constructors

FullySpecified

SheetOnly

Unspecified

data Lit

types of literals

Constructors

Bool

Int

data LitOf

values of literals

Constructors

BoolOf !Bool

IntOf !Int

data TensorDescriptor

a tensor (type) is described by the number of dimensions of each component

Constructors

TensorDescriptor

Fields

base :: Term Infer

dims :: [Term Check]

data Term (m :: Mode) where

Constructors

Inf :: !(Term Infer) -> Term Check
inferrable terms are checkable

Lam :: !Arg -> !(Maybe CaseInsensitive) -> !(Term Check) -> Term Check
lambda abstraction

Ann :: !(Term Check) -> !(Term Infer) -> Term Infer
annotated terms e: t

Set :: !Int -> Term Infer
hierarchy of type universes

Pi :: !Arg -> !(Maybe CaseInsensitive) -> !(Term Check) -> !(Term Check) -> Term Infer
dependent function (Π)

Bound :: !Int -> Term Infer
bound variable

Free :: !Name -> Term Infer
free variable

Ref :: !(URI, Text) -> RefInfo -> CellRange -> Term Infer
cell references

Lit :: !Lit -> Term Infer
literal types

LitOf :: !LitOf -> Term Infer
literal terms

Tensor :: !TensorDescriptor -> Term Infer
tensor type (only for quoting)

TensorOf :: !TensorDescriptor -> !(Array (Term Check)) -> Term Infer
tensor value (represented as a multi-dimensional array, only for quoting)

App :: !(Term Infer) -> !(Arg, Term Check) -> Term Infer
application

Elaborate :: !(Term Check) -> Term Infer
elaborated terms (do not use before typechecking!)

pattern (:$) :: Term Infer -> Term Check -> Term Infer

elaborate :: Term Check -> Term Infer

smart constructor (x when passing Elaborate (Inf x))

boolOf :: Bool -> Term Infer

intOf :: Int -> Term Infer

lam :: Arg -> CaseInsensitive -> Term Check -> Term Check

splitApp :: Term Infer -> (Term Infer, [(Arg, Term Check)])

Substitutions

subst :: Int -> Term Infer -> Term m -> Term m

substTensor :: Int -> Term Infer -> TensorDescriptor -> TensorDescriptor

braced :: [Doc ann] -> Doc ann

showExcel26 :: CellAddr -> String

prettyURI :: URI -> Doc ann

prettyURI' :: URI -> Text

escapeUri :: Text -> Doc ann

genFresh :: [CaseInsensitive] -> CaseInsensitive

sub :: Int -> Int -> Term m -> Term m

subTensor :: Int -> Int -> TensorDescriptor -> TensorDescriptor