Recalc.Engine

Description

The evaluation strategy is inspired by the paper "Build systems à la carte", it maintains a DocumentStore that tracks types, values and meta data for all cells.

The implementation maintains a DocumentStore that tracks types, values and meta data for all cells.

References: Andrey Mokhov, Neil Mitchell and Simon Peyton Jones. Build Systems à la Carte. Proceedings of the ACM on Programming Languages, Volume 2, Issue ICFP.

Cell-IDs

type SheetId = (URI, Text)

Spreadsheets are uniquely determined by a resource id and sheet name

data URI

Represents a general universal resource identifier using its component parts.

For example, for the URI

  foo://anonymous@www.haskell.org:42/ghc?query#frag

the components are:

data Text

A space efficient, packed, unboxed Unicode text type.

type CellAddr = (Int, Int)

Row and column (both zero-indexed)

type CellRange = (CellAddr, CellAddr)

Beginning (top left) and end (bottom right) of a range

Generic spreadsheet language interface

class Language term where

The spreadsheet language abstraction

Associated Types

type EnvOf term

type ErrorOf term

type ValueOf term

Methods

deps :: term -> Set CellRange

Compute dependencies of a term

newEnv :: SheetId -> EnvOf term

infer :: term -> Fetch (EnvOf term) (ErrorOf term) (ValueOf term) (ValueOf term)

All terms can be evaluated in the Fetch context (a dependently typed language is assumed: values and types coincide)

inferValue :: EnvOf term -> ValueOf term -> Maybe (ValueOf term)

All values can be inferred

eval :: term -> Fetch (EnvOf term) (ErrorOf term) (ValueOf term) (ValueOf term)

All terms can be evaluated in the Fetch context

data FetchError err

Evaluation of cells can always fail due to invalid formulas or refs

Constructors

InvalidFormula (ParseErrorBundle String Void)

RefError

OtherError err

errorTitle :: (a -> Text) -> FetchError a -> Text

class Isn't cell where

Typeclass for values that "default-when-absent". For example, Bool when not given should be "the same" as False. Used for meta-data only.

Methods

isn't :: cell -> Bool

class Meta cell where

can be merged (à la "missing data kept, null removes it")

Methods

merge :: cell -> cell -> cell

data Ix

The spreadsheet engine can be queried for re-evaluation of cells (types and values) and volatile results.

Constructors

Cell Kind SheetId CellAddr

Volatile

type Fetch term a = Fetch (EnvOf term) (ErrorOf term) (ValueOf term) a

type TypeAnnotation dat = Maybe (ValueOf (TermOf dat))

type Engine doc sheet cell term = EngineT doc sheet cell term Identity

type EngineState = EngineState Slow

type EngineT doc sheet cell term f = EngineT Slow doc sheet cell term f

class Language (TermOf dat) => Input dat where

input data can be arbitrary

Associated Types

type TermOf dat

type MetaOf dat

Methods

termOrValueOf :: SheetId -> CellAddr -> dat -> Maybe (Parsed (Either (TermOf dat) (ValueOf (TermOf dat))))

metaOf :: dat -> MetaOf dat

type Parsed = Either (ParseErrorBundle String Void)

type MetaChangesOf dat = [(CellAddr, MetaOf dat)]

meta changes which don't map to terms/values, do not affect re-computation

type ChangesOf dat = ([CellAddr], [(CellAddr, MetaOf dat, ParseErrorBundle String Void)], [(CellAddr, MetaOf dat, ValueOf (TermOf dat))], [(CellAddr, MetaOf dat, TermOf dat)])

validated cells, all changes affect recomputation (deletions, parse errors, new values, new terms)

type NewValues dat = [((SheetId, CellAddr), Either (FetchErrorOf (TermOf dat)) (ValueOf (TermOf dat), TypeAnnotation dat))]

recomputing will return new (value, type)s per cell (or an error)

runEngineT :: EngineT doc sheet cell term f a -> EngineState doc sheet cell term -> f (a, EngineState doc sheet cell term)

updateMeta :: (Isn't cell, Meta cell, Monad f) => SheetId -> [(CellAddr, cell)] -> EngineT doc sheet cell term f ()

validateCells :: Input dat => SheetId -> [(CellAddr, dat)] -> (MetaChangesOf dat, ChangesOf dat)

recompute :: forall dat doc sheet f. (Monad f, Show (ValueOf (TermOf dat)), Input dat, Isn't (MetaOf dat)) => SheetId -> ChangesOf dat -> EngineT doc sheet (MetaOf dat) (TermOf dat) f (Either [(SheetId, CellAddr)] (NewValues dat))

recompute new store from validated new inputs

type DS doc sheet cell term = DocumentStore doc sheet cell term (ValueOf term) (FetchErrorOf term)

newEngineState :: DependencyMap dm => EngineState dm doc sheet cell term

modifyDocs :: Monad f => (DS doc sheet cell term -> DS doc sheet cell term) -> EngineT dm doc sheet cell term f ()

insertDocument :: URI -> doc -> DocumentStore doc sheet cell term value err -> DocumentStore doc sheet cell term value err

insertSheet :: SheetId -> sheet -> DocumentStore doc sheet cell term value err -> DocumentStore doc sheet cell term value err

deleteSheet :: SheetId -> DocumentStore doc sheet cell term value err -> DocumentStore doc sheet cell term value err

updateDocument :: URI -> (doc -> doc) -> DocumentStore doc sheet cell term value err -> DocumentStore doc sheet cell term value err

updateSheet :: SheetId -> (sheet -> sheet) -> DocumentStore doc sheet cell term value err -> DocumentStore doc sheet cell term value err

fetchType :: SheetId -> CellAddr -> Fetch env err value value

fetchValue :: SheetId -> CellAddr -> Fetch env err value value

getEnv :: Fetch env err value env

localEnv :: (env -> env) -> Fetch env err value a -> Fetch env err value a