Recalc.Syntax.Term

Description

Types definitions for a very basic dependently typed lambda calculus based on "A tutorial implementation of a dependently typed lambda calculus".

Terms

data Mode

Constructors

Check

Infer

Names

newtype CaseInsensitive

Constructors

CaseInsensitive

Fields

originalText :: Text

caseInsensitive :: CaseInsensitive -> Text

data Name

Constructors

Global !CaseInsensitive

Local !(Maybe CaseInsensitive) !Int
when passing binder, bound names are converted to locally free ones

Quote !Int
same but during quoting

pat :: Text -> Maybe CaseInsensitive

data RefInfo

cell references may be given [file_name.rc]sheet!C3, sheet!C3, or C3; corresponding to FullySpecified, SheetOnly, and Unspecified respectively

Constructors

FullySpecified

SheetOnly

Unspecified

data Lit

types of literals

Constructors

Bool

Int

data LitOf

values of literals

Constructors

BoolOf !Bool

IntOf !Int

data TensorDescriptor

a tensor (type) is described by the number of dimensions of each component

Constructors

TensorDescriptor

Fields

base :: Term Infer

dims :: [Int]

data Term (m :: Mode) where

Constructors

Inf :: !(Term Infer) -> Term Check
inferrable terms are checkable

Lam :: !(Maybe CaseInsensitive) -> !(Term Check) -> Term Check
lambda abstraction

Ann :: !(Term Check) -> !(Term Infer) -> Term Infer
annotated terms e: t

Set :: !Int -> Term Infer
hierarchy of type universes

Pi :: !(Maybe CaseInsensitive) -> !(Term Check) -> !(Term Check) -> Term Infer
dependent function (Π)

Bound :: !Int -> Term Infer
bound variable

Free :: !Name -> Term Infer
free variable

Ref :: !(URI, Text) -> RefInfo -> CellRange -> Term Infer
cell references

Lit :: !Lit -> Term Infer
literal types

LitOf :: !LitOf -> Term Infer
literal terms

Tensor :: !TensorDescriptor -> Term Infer
tensor type (only for quoting)

TensorOf :: !TensorDescriptor -> !(Array (Term Check)) -> Term Infer
tensor value (represented as a multi-dimensional array, only for quoting)

(:$) :: !(Term Infer) -> !(Term Check) -> Term Infer
application

boolOf :: Bool -> Term Infer

intOf :: Int -> Term Infer

lam :: CaseInsensitive -> Term Check -> Term Check

splitApp :: Term Infer -> (Term Infer, [Term Check])

Substitutions

subst :: Int -> Term Infer -> Term m -> Term m

substTensor :: Int -> Term Infer -> TensorDescriptor -> TensorDescriptor

braced :: [Doc ann] -> Doc ann

showExcel26 :: CellAddr -> String

prettyURI :: URI -> Doc ann

prettyURI' :: URI -> Text

escapeUri :: Text -> Doc ann

genFresh :: [CaseInsensitive] -> CaseInsensitive

splitFun :: Term Infer -> ([Term Check], Term Check)